requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
sqlalchemy==2.0.20
pydantic==1.10.9
python-dotenv==1.0.0
pytest==7.4.0
pyodbc==4.0.35
email-validator==1.3.1
passlib==1.7.4
python-jose==3.4.0
bcrypt==4.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
google-auth-oauthlib==1.2.1
jinja2==3.1.2
python-multipart==0.0.6